Controlling and Users done
Fixes missing

+++ Model/OvertimeDetail.cs
@@ -6,6 +6,6 @@ namespace TimeManager.Model
   {
     public Guid ID { get; set; }
     public String Description { get; set; }
-    public Int32 Rate { get; set; }
+    public Decimal Rate { get; set; }
   }
 }
